+++ src/routes/index.ts
@@ -1,0 +1,7 @@
+import Root from "./root";
+
+const exportedRoutes = {
+  Root: Root,
+};
+
+export default exportedRoutes;